+++ lib/src/yeeter_delegate.dart
@@ -8,18 +8,17 @@ import 'yeet.dart';
 
 final GlobalKey<NavigatorState> _navigatorKey = GlobalKey<NavigatorState>();
 final _heroController = HeroController();
+List<Page> _pages = [];
 
 /// Put this as your routerDelegate in [MaterialApp.router].
-class YeeterDelegate extends RouterDelegate<RouteInformation>
-    with ChangeNotifier {
+
+class YeeterDelegate extends RouterDelegate<String> with ChangeNotifier {
   final Yeet _yeet;
-  List<Page> _pages;
 
   YeeterDelegate({
     required Yeet yeet,
-  })   : _yeet = yeet,
-        _pages = [] {
-    this.yeet('/');
+  }) : _yeet = yeet {
+    this.yeet(currentConfiguration);
   }
 
   List<Page>? _dfs(
@@ -30,17 +29,18 @@ class YeeterDelegate extends RouterDelegate<RouteInformation>
     Map<String, String> query,
   ) {
     final pages = <Page>[];
-
     if (node.regExp != null) {
       // Handling relative and non-relative paths correctly
       final isRootPath = node.path!.startsWith('/');
       if (isRootPath) matchedTill = 0;
       final match = node.regExp!.matchAsPrefix(path.substring(matchedTill));
-
       if (match != null) {
+        final isFinal = matchedTill + match.end == path.length;
         params.addAll(extract(node.parameters, match));
         if (node.builder != null) {
-          final key = ValueKey(path.substring(0, matchedTill + match.end));
+          final queryPath = Uri(queryParameters: query).query;
+          final key = ValueKey(path.substring(0, matchedTill + match.end) +
+              (isFinal && queryPath.isNotEmpty ? '?$queryPath' : ''));
           final child = node.builder!(params, query);
           if (node.transitionsBuilder == null) {
             if (UniversalPlatform.isIOS || UniversalPlatform.isMacOS) {
@@ -74,11 +74,13 @@ class YeeterDelegate extends RouterDelegate<RouteInformation>
             ));
           }
         }
-        if (matchedTill + match.end == path.length) {
+        if (isFinal) {
           // The matching is final.
           return pages;
         }
-        if (path[matchedTill + match.end] == '/') {
+        if (path[matchedTill + match.end - 1] == '/') {
+          matchedTill += match.end;
+        } else if (path[matchedTill + match.end] == '/') {
           matchedTill += match.end + 1;
         } else {
           matchedTill = 0;
@@ -138,7 +140,7 @@ class YeeterDelegate extends RouterDelegate<RouteInformation>
       )!;
       notifyListeners();
     } else {
-      final location = currentConfiguration!.location!;
+      final location = Uri.parse(currentConfiguration).path;
       yeet(location + (location != '/' ? '/' : '') + path);
     }
   }
@@ -155,14 +157,13 @@ class YeeterDelegate extends RouterDelegate<RouteInformation>
   }
 
   @override
-  Future<void> setNewRoutePath(RouteInformation configuration) {
+  Future<void> setNewRoutePath(String path) {
     return Future.sync(() {
-      yeet(configuration.location!);
+      yeet(path);
     });
   }
 
   @override
-  RouteInformation? get currentConfiguration => _pages.isNotEmpty
-      ? RouteInformation(location: (_pages.last.key as ValueKey).value)
-      : null;
+  String get currentConfiguration =>
+      _pages.isNotEmpty ? (_pages.last.key as ValueKey).value : '/';
 }